pyproject.toml
[tool.poetry]
name = "genai-rag-platform"
version = "0.1.0"
description = "GenAI platform with PDF ingestion, vector search (Qdrant), graph search (Neo4j), and LangGraph chat API."
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "0.115.0"
langchain = "~0.3.0"
langchain-community = "~0.3.0"
langchain-openai = "~0.2.0"
langgraph = "~0.2.0"
neo4j = "~5.0"
pypdf = "~5.0"
python-dotenv = "~1.0.0"
qdrant-client = ">=1.9.0,<2.0.0"
sentence-transformers = "~3.0"
uvicorn = "~0.30.0"

[tool.poetry.group.dev.dependencies]
mypy = ">=1.0.0"
ruff = ">=0.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true

[tool.ruff]
target-version = "py310"
line-length = 100
exclude = [".venv", "__pycache__", ".git"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["app", "api", "llm", "monitoring"]
